refactor: Extracted function (updateReferences)
