+++ src/steps/create-signatures.ts
@@ -1,8 +1,6 @@
-/* eslint-disable @typescript-eslint/ban-ts-comment */
 import { readFileSync } from 'node:fs';
 import { join } from 'node:path';
 
-import { AST } from '@codemod-utils/ast-javascript';
 import { createFiles } from '@codemod-utils/files';
 
 import type { Context, Options } from '../types/index.js';
@@ -12,10 +10,9 @@ import {
   transformEntityName,
 } from '../utils/files.js';
 import {
-  convertArgsToSignature,
   getBaseComponentName,
-  isSignature,
   passSignatureToBaseComponent,
+  updateReferences,
 } from './create-signatures/index.js';
 
 type Data = {
@@ -29,7 +26,8 @@ function createSignature(file: string, data: Data): string {
     return file;
   }
 
-  const { interfaceName, newFile } = passSignatureToBaseComponent(file, {
+  // eslint-disable-next-line prefer-const
+  let { interfaceName, newFile } = passSignatureToBaseComponent(file, {
     baseComponentName,
     data,
   });
@@ -38,77 +36,12 @@ function createSignature(file: string, data: Data): string {
     return newFile;
   }
 
-  const traverse = AST.traverse(true);
+  ({ newFile } = updateReferences(newFile, {
+    data,
+    interfaceName,
+  }));
 
-  const ast = traverse(newFile, {
-    visitTSInterfaceDeclaration(path) {
-      // @ts-ignore: Assume that types from external packages are correct
-      if (path.node.id.name !== interfaceName) {
-        return false;
-      }
-
-      // @ts-ignore: Assume that types from external packages are correct
-      path.node.id.name = `${data.entity.classifiedName}Signature`;
-
-      const typeParameter = path.node.body;
-      if (isSignature(typeParameter.body)) {
-        return false;
-      }
-
-      typeParameter.body = convertArgsToSignature({
-        b: AST.builders,
-        nodes: typeParameter.body,
-      });
-
-      return false;
-    },
-
-    visitTSTypeAliasDeclaration(path) {
-      // @ts-ignore: Assume that types from external packages are correct
-      if (path.node.id.name !== interfaceName) {
-        return false;
-      }
-
-      // @ts-ignore: Assume that types from external packages are correct
-      path.node.id.name = `${data.entity.classifiedName}Signature`;
-
-      const typeParameter = path.node.typeAnnotation;
-
-      // @ts-ignore: Assume that types from external packages are correct
-      if (isSignature(typeParameter.members)) {
-        return false;
-      }
-
-      // @ts-ignore: Assume that types from external packages are correct
-      typeParameter.members = convertArgsToSignature({
-        b: AST.builders,
-        // @ts-ignore: Assume that types from external packages are correct
-        nodes: typeParameter.members,
-      });
-
-      return false;
-    },
-
-    visitTSTypeReference(path) {
-      if (path.node.typeName.type !== 'Identifier') {
-        return false;
-      }
-
-      if (path.node.typeName.name !== interfaceName) {
-        return false;
-      }
-
-      if (path.node.typeName.name.endsWith('Signature')) {
-        return false;
-      }
-
-      path.node.typeName.name = `${data.entity.classifiedName}Signature['Args']`;
-
-      return false;
-    },
-  });
-
-  return AST.print(ast);
+  return newFile;
 }
 
 export function createSignatures(context: Context, options: Options): void {
