+++ src/steps/create-signatures/index.ts
@@ -1,4 +1,3 @@
-export * from './convert-args-to-signature.js';
 export * from './get-base-component-name.js';
-export * from './is-signature.js';
 export * from './pass-signature-to-base-component.js';
+export * from './update-references.js';
